Web App/Back End/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
scikit-learn==1.6.1
networkx==3.3
node2vec==0.5.0
torch==2.7.0
torch-geometric==2.6.0
transformers==4.46.2
sentence-transformers==3.2.1
rdkit==2024.9.3
requests==2.32.3
Werkzeug==3.1.3
gunicorn==21.2.0
